+++ src/main/java/com/erebelo/springh2demo/repository/ProductOrderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductOrderRepository extends JpaRepository<ProductOrderEntity, Integer> {
+public interface ProductOrderRepository extends JpaRepository<ProductOrderEntity, Long> {
 
 }
